requirements.txt
# Neo4j database
neo4j==5.14.1

# Data processing
pandas==2.1.3
numpy>=1.21.0

# Visualization
networkx==3.2.1
matplotlib==3.8.2

# Machine Learning & NLP
scikit-learn>=1.0.0
spacy>=3.7.2
joblib>=1.2.0

# Deep Learning (PyTorch)
torch>=2.0.0

# Transformers & LLM
transformers>=4.30.0
tokenizers>=0.13.0
datasets>=2.12.0
accelerate>=0.20.0
sentencepiece>=0.1.99

# FastText for embeddings
fasttext>=0.9.2

# RAG (Retrieval-Augmented Generation)
faiss-cpu>=1.7.0  # or faiss-gpu for GPU support
sentence-transformers>=2.0.0
openai>=1.0.0

# Utilities
tqdm>=4.65.0
